requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
defusedxml==0.7.1
Django==3.0.14
djangorestframework==3.11.2
httpie==3.2.1
idna==3.3
multidict==6.0.2
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
sqlparse==0.4.2
urllib3==1.26.9
